# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/payment/{id}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get payment by ID",
                "operationId": "4e9439126d32d3164ef4295c155c0b82",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of the payment",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "123e4567-e89b-12d3-a456-426614174000"
                                                },
                                                "name_client": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "cpf": {
                                                    "type": "string",
                                                    "example": "12345678901"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Payment for service"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 150.75
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "method_payment": {
                                                    "type": "string",
                                                    "example": "credit_card"
                                                },
                                                "paid_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-05-30 22:00:00"
                                                },
                                                "merchant": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 8
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Teste User"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "rp@telnav.com"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "merchant"
                                                        },
                                                        "balance": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 31.02
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": null,
                                                            "nullable": true
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "example": null,
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}